Degrader/requirements.txt
PyQt6
numpy
scipy
Pillow